pyproject.toml
[project]
name = "ipq"
version = "0.1.0"
description = "Fast IP, CIDR, and domain information query tool for network operations"
readme = "README.md"
license = "CC0-1.0"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.6",
    "ipwhois>=1.3",
    "httpx>=0.27",
    "typer>=0.12",
    "rich>=13.0",
]

# AIDEV-NOTE: Traceroute now uses raw ICMP sockets directly (no icmplib needed).
# Requires root or CAP_NET_RAW for the fast parallel engine; falls back to
# system traceroute/tracepath when unprivileged.

[project.urls]
Homepage = "https://github.com/linsomniac/ipq"
Repository = "https://github.com/linsomniac/ipq"

[project.scripts]
ipq = "ipq.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ipq"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
